fix: fix bundler identifier & package name
