+++ android/app/build.gradle
@@ -136,7 +136,7 @@ android {
         implementation 'com.android.support:multidex:1.0.3'
     }
     defaultConfig {
-        applicationId "com.dingureureactnative"
+        applicationId "com.gb.dingureureactnative"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
