+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">dingureuReactNative</string>
+    <string name="app_name">딩굴</string>
 </resources>
