+++ app.json
@@ -1,4 +1,4 @@
 {
   "name": "dingureuReactNative",
-  "displayName": "dingureuReactNative"
+  "displayName": "딩굴"
 }
