+++ src/api/articles/index.ts
@@ -1,0 +1,9 @@
+import { api } from "../common"
+
+export const articleLocal = () => {
+    return api.get('/articles/local/')
+}
+
+export const articleSchool = () => {
+    return api.get('/articles/school/')
+}
